backend/requirements.txt
Flask
flask-cors
flask-jwt-extended
pymongo
bcrypt
python-dotenv
pandas
itsdangerous
gunicorn

